feat: change default LICENSE file name from LICENSE.txt to LICENSE
After a short survey this seems to be the most used file name for the license. Since this package isn't ver popular yet, introducing this breaking change shouldn't be much of a problem.

BREAKING CHANGE: The default license file is now called LICENSE instead of LICENSE.txt
If you want to use the old file name, just set it in the options.