feat(badges): show build status only for main branch in travis badge
Previously the build status of the latest branch was shown. This could
lead to confusion if the latest branch was a breaking PR while the main
branch of the repo built fine. Show the build status of the main branch
instead.

BREAKING CHANGE: Previously the travis badge showed the build status of
the newest commit. It now shows the build status of the newest commit
from the main branch, by default master. You can change the main branch
in the options. It is not possible anymore to show the build status
across all branches.