feat(docs): use GitHub Pages instead of rawgit to serve docs
rawgit has been discontinued, but GitHub Pages helps to do the same as
long as docs are in ./docs

BREAKING CHANGE: Instead of gitraw now GitHub Pages is used to deliver
documentation. To use this you have to save your docs in ./docs and
enable master/docs as the location for GitHub Pages.